Added a PlotMap
This singleton will keep track of the pipelines created for a given Plot
and set of Variables so that we can display overlays correctly even if
pipelines get reexecuted/moved/copied or changes happen in VisTrails.

At some point, this map should get serialized to the VT file so that the
Plot and Variables used to create the pipeline can still be displayed or
changed after reloading from the file.