Wrote an AdvancedLineEdit component
Renamed dat.gui.lists to dat.gui.generic.

Wrote an AdvancedLineEdit component based on QLineEdit that adds a
"reset to default" behavior and a validation system, with color
feedback.

This new component is used in LoadVariableDialog for the variable name
so that it can be reset to its default value and display whether it is
currently using default value (blue) and whether the entered name is
acceptable (doesn't conflict with existing variables; green/red).