Rewrites the Manager/PlotMap classes
There is now a dat.global_data:GlobalManager singleton that holds the
Plots and VariableLoaders. It is similar to the old Manager, except it
doesn't have the Variables anymore.

There is now a dat.vistrail_data:VistrailManager singleton that holds
the Variables and pipeline -> recipe mapping through VistrailData
objects, like the Variable part of the old Manager + the old PlotMap.

This allows multiple vistrail to be managed at the same time, each with
its own set of Variables and pipeline -> recipe mappings.