Fixes setWidget not getting called bug
If the spreadsheet receives a DisplayCellEvent, setCellByType() is
called; there, if the new cell type is the same as the previous one,
updateContents() is called on the cell, and the container doesn't get
notified.

This has been rectified with the introduction of
CellContainerInterface#contentsUpdated(). In DAT, we use it to retrieve
the new pipelineInfo and update the recipe, like we did previously in
setWidget().