Config widget is now an overlay.
Devs can provide their own plot config widget by passing in the
widget class when defining plots in the package. The custom plot
widget must subclass dat.gui.overlays.PlotConfigOverlay, take
cellcontainer as the first argument to __init__ and use it to
call Overlay.__init__, and must override setup(self, cell, plot)
which is called before the plot config is shown.

Work needs to be done related to updating the plot while the overlay
is displayed. Currently it will always remove the overlay. There also
seems to be an issue on Mac where the cell corner button to display
the variable overlay doesn't show up.