Fixes a potential bug
add_function() was using module.module_descriptor.get_port_spec(),
contrary to select_output_port which used module.get_port_spec().

The difference is that a module can have specific port specs not
declared in the ModuleDescriptor, and these are checked as well when
using Module#get_port_spec(). This is what we want.