Displays the overlay for non-executable pipelines
When restoring a vistrail from a file, the pipeline might not be
executable (because of missing parameters) or might cause an error. In
this case, we still want to display the overlay for that version in the
cell.

dat/vistrail_data.py:
  - _discover_cell_pipelines() fills in _cell_location and
    _cell_to_version from the CellLocation modules in the pipeline. This
    logic was previously in Application.

dat/gui/application.py:
  - Get the cell/pipeline couples from vistrail_data
  - Assign (through setCellWidget()) a DATCellContainer (with error
    optionally set)
  - setCellWidget() had a bug, solved in the submodule

dat/gui/cellcontainer.py:
  - DATCellContainer's constructor now accepts an 'error' argument.