Works around an issue in paramexplore
VisTrails's parameter view tracks the current controller and pipeline
separately but expects the pipeline to refer to the controller. This
could be wrong as DAT changes the pipeline (when executing) in its
handler for 'controller_changed', causing new 'pipeline_changed'
notifications to be emitted before other handlers for
'controller_changed' had a chance to run.

We just defer the execution of visualizations using Qt's meta-object
method invocation system, to let other handlers for 'controller_changed'
run first.