Puts the show/hide overlay buttons on the toolbar
This refactors cellcontainer.py quite a bit.

DATCellContainer no longer inherit QCellContainer, but
CellContainerInterface and QWidget directly. This simplifies the
problems that previously arose from having no way of doing stuff between
the Qt component's creation and the call to setWidget() by the base
class.
We also get rid of the layout that wasn't so useful.

Removes the _show_button and _hide_button QPushButton's.

Implements the containerToolBar() method.

This might cause issues with error reporting (the error icon).