Reads ConstantPort's from aliased parameters
It is now possible to add constant parameters by simply adding an alias
to a parameter on that port. The alias will be used as name of the
constant port, and the type and default value of the port will be used.

TODO: we should use the value set on the port as default value.