Remove detection of pkg when constructing Variable
When a Variable is created, it looks at the calling package to try and
set _vt_package_id, used to resolve short package names. However this
only worked for __init__ and init modules directly at the top of a
package.

It doesn't seem such a good idea, so this removes it. It could
alternatively be fixed to work from any module under the package.